# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MeResponse": {
        "properties": {
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_login_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Login At"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "title": "Permissions",
            "type": "array"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "title": "Roles",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/UserStatus"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "avatar_url",
          "status",
          "roles",
          "last_login_at",
          "created_at",
          "permissions"
        ],
        "title": "MeResponse",
        "type": "object"
      },
...
      "ProfileOut": {
        "properties": {
          "bio": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bio"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "has_resume": {
            "title": "Has Resume",
            "type": "boolean"
          },
          "headline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headline"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "onboarding_completed": {
            "title": "Onboarding Completed",
            "type": "boolean"
          },
          "preferred_locations": {
            "items": {
              "type": "string"
            },
            "title": "Preferred Locations",
            "type": "array"
          },
          "remote_preference": {
            "$ref": "#/components/schemas/RemotePreference"
          },
          "target_roles": {
            "items": {
              "type": "string"
            },
            "title": "Target Roles",
            "type": "array"
          },
          "years_experience": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Years Experience"
          }
        },
        "required": [
          "name",
          "email",
          "headline",
          "bio",
          "years_experience",
          "target_roles",
          "preferred_locations",
          "remote_preference",
          "onboarding_completed",
          "has_resume"
        ],
        "title": "ProfileOut",
        "type": "object"
      },
      "ProfileUpdate": {
        "properties": {
          "bio": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bio"
          },
          "complete_onboarding": {
            "default": false,
            "title": "Complete Onboarding",
            "type": "boolean"
          },
          "headline": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headline"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "preferred_locations": {
            "items": {
              "type": "string"
            },
            "maxItems": 10,
            "title": "Preferred Locations",
            "type": "array"
          },
          "remote_preference": {
            "$ref": "#/components/schemas/RemotePreference",
            "default": "any"
          },
          "target_roles": {
            "items": {
              "type": "string"
            },
            "maxItems": 10,
            "title": "Target Roles",
            "type": "array"
          },
          "years_experience": {
            "anyOf": [
              {
                "maximum": 60.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Years Experience"
          }
        },
        "title": "ProfileUpdate",
        "type": "object"
      },
...
      "RemotePreference": {
        "enum": [
          "remote",
          "hybrid",
          "onsite",
          "any"
        ],
        "title": "RemotePreference",
        "type": "string"
      },
...
      "UserStatus": {
        "enum": [
          "active",
          "suspended"
        ],
        "title": "UserStatus",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "GlideUp access token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Find jobs. Practice interviews. Get hired.",
    "title": "GlideUp API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/users/me": {
      "get": {
        "operationId": "read_me_api_v1_users_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Me",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/me/profile": {
      "get": {
        "operationId": "read_profile_api_v1_users_me_profile_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Profile",
        "tags": [
          "users"
        ]
      },
      "put": {
        "operationId": "update_profile_api_v1_users_me_profile_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Profile",
        "tags": [
          "users"
        ]
      }
    },
...
  }
}